8 - Debugging & Packs
composer require debug --dev

That command installed & severals features in one package.

The dark side of packages is that we can't control the version of a librarie or delete an unused one.
To resolve this problem we can just unpack the pack.

composer unpack debug.

We installed phpunit pack too.

composer require phpunit --dev