57 - Collection Magic with Criteria
Doctrine has a super powerful and amazing feature called Criteria.

Instead of looping over all the data, add $criteria = Criteria - the one from Doctrine - Criteria::create().

The Criteria object is similar to the QueryBuilder we're used to, but with a slightly different, well, slightly more confusing syntax.

By using Criteria, we get super efficient filtering.

We can also re-use Criteria in a QueryBuilder because of a static call.